lesson_1_6 completed + file structure refactored. Reusabe Input with Label Component & Component Composition. Auto-Focus with ref. Remove button.
